refactor(analysis): externalize abstract_types inline tests
