+++ foundation/analysis/src/abstract_types/mod.rs
@@ -5,6 +5,11 @@
 pub mod angle;
 pub mod scalar;
 
+#[cfg(test)]
+pub mod angle_tests;
+#[cfg(test)]
+pub mod scalar_tests;
+
 // 主要な型とトレイトを再エクスポート
 pub use angle::{Angle, AngleType};
 pub use scalar::Scalar;
